remove editor options as yawgmoth is only a viewer
