+++ deck-editor/src/app/components/decklist/controller.js
@@ -2,27 +2,28 @@ import angular from 'angular';
 import template from './template.html';
 
 class DeckListCtrl {
-  constructor($scope, deckService, cardService) {
-    this.$scope = $scope;
-    this.deckService = deckService;
-    this.cardService = cardService;
-  }
+	constructor($scope, $stateParams, deckService, cardService) {
+		this.$scope = $scope;
+		this.deckService = deckService;
+		this.cardService = cardService;
+		this.deckIndex = $stateParams.deckIndex;
+	}
 
-  $onInit() {
-    this.deckService.getDecks().then(decks => {
-      let deck = decks[0];
-      this.cardService.getCardsFromText(deck.cards).then(cards => {
-        this.deck = {name: deck.name, cards: cards};
-      });
-    });
-  }
+	$onInit() {
+		this.deckService.getDecks().then(decks => {
+			let deck = decks[this.deckIndex];
+			this.cardService.getCardsFromText(deck.cards).then(cards => {
+				this.deck = {name: deck.name, cards: cards};
+			});
+		});
+	}
 }
 
 export default angular.module('ksDeckList', [])
-  .controller('DeckListCtrl', ['$scope', 'deckService', 'cardService', DeckListCtrl])
-  .component('ksDeckList', {
-    template: template,
-    controllerAs: 'ctrl',
-    controller: 'DeckListCtrl'
-  })
-  .name;
+	.controller('DeckListCtrl', ['$scope', '$stateParams', 'deckService', 'cardService', DeckListCtrl])
+	.component('ksDeckList', {
+		template: template,
+		controllerAs: 'ctrl',
+		controller: 'DeckListCtrl'
+	})
+	.name;
