+++ deck-editor/src/app/components/decklist/template.html
@@ -1,4 +1,4 @@
-<h1 class="ui header">Visual Deck List</h1>
+<h1 class="ui header">{{ctrl.deck.name}}</h1>
 <div id="deckList" class="visual-deck-list" ng-if="ctrl.deck">
     <div class="deck-name">
         <div>{{ctrl.deck.name}}</div>
