+++ deck-editor/src/app/components/deckshot/template.html
@@ -1,4 +1,3 @@
-<h1 class="ui header">Snapshot</h1>
 <div id="deckCapture"></div>
 <div ng-if="ctrl.snapshotCreated">
   <p>* right click on the snapshot and click "save image as ..." to download the image.</p>
