+++ deck-editor/src/app/views/deckDetail/template.html
@@ -1,0 +1,5 @@
+<div class="ui container segment">
+	<ks-deck-list2></ks-deck-list2>
+	<br/>
+	<a class="ui button" ui-sref="viewer({deckIndex: ctrl.deckIndex})">visual view</a>
+</div>
