+++ deck-editor/src/app/views/deckList/controller.js
