--- deck-editor/src/app/views/editor/template.html
@@ -1,14 +1,0 @@
-<div class="ui container segment">
-    <div class="ui grid stackable">
-        <div class="eight wide column">
-            <ks-deck-form></ks-deck-form>
-            <br/>
-            <button class="ui button primary fluid" ng-click="ctrl.updatePreview()">update preview</button>
-            <br/>
-            <a class="ui button fluid" ui-sref="viewer({deckIndex: ctrl.deckIndex})">visual view</a>
-        </div>
-        <div class="eight wide column">
-            <ks-deck-list2></ks-deck-list2>
-        </div>
-    </div>
-</div>
