--- deck-editor/src/app/views/viewer/template.html
@@ -1,11 +1,0 @@
-<div class="ui container segment">
-    <div class="ui grid stackable">
-        <div class="eight wide column">
-            <ks-deck-list></ks-deck-list>
-            <button class="ui button primary fluid" ng-click="ctrl.createSnapshot()">create snapshot</button>
-        </div>
-        <div class="eight wide column">
-            <ks-deck-shot></ks-deck-shot>
-        </div>
-    </div>
-</div>
