+++ deck-editor/src/app/views/visualViewer/template.html
@@ -1,0 +1,11 @@
+<div class="ui container segment">
+	<div class="ui grid stackable">
+		<div class="eight wide column">
+			<ks-deck-list></ks-deck-list>
+			<button class="ui button primary" ng-click="ctrl.createSnapshot()">create snapshot</button>
+		</div>
+		<div class="eight wide column">
+			<ks-deck-shot></ks-deck-shot>
+		</div>
+	</div>
+</div>
